Extract an operations layer from rootstock/commands/
Fixes #111. Build/add/refresh orchestration moves to
rootstock/operations.py: keyword-only functions returning structured
results (InstallResult, AddResult), progress via callback, failures as
OperationError. Commands become thin argparse adapters; serve.py no
longer imports add.py's privates; external drivers can install envs
and add checkpoints without shelling out to the CLI.
