+++ tests/commands/test_install_spec.py
@@ -15,7 +15,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from rootstock.commands.install import (
+from rootstock.operations import (
     ROOTSTOCK_GITHUB_URL,
     _rootstock_install_spec,
 )
@@ -111,33 +111,28 @@ def test_install_command_passes_helper_spec_as_final_arg(tmp_path, monkeypatch, 
         result.stdout = ""
         return result
 
-    monkeypatch.setattr("rootstock.commands.install.subprocess.run", fake_run)
-    monkeypatch.setattr("rootstock.commands.install.shutil.copy", lambda *a, **k: None)
+    monkeypatch.setattr("rootstock.operations.subprocess.run", fake_run)
+    monkeypatch.setattr("rootstock.operations.shutil.copy", lambda *a, **k: None)
     # No vendored wheel available: force the index-spec fallback (also keeps
     # the test off the network).
+    monkeypatch.setattr("rootstock.operations._vendor_rootstock_wheel", lambda root: None)
+    monkeypatch.setattr("rootstock.operations._precompile_environment", lambda *a, **k: None)
     monkeypatch.setattr(
-        "rootstock.commands.install._vendor_rootstock_wheel", lambda root: None
-    )
-    monkeypatch.setattr(
-        "rootstock.commands.install._precompile_environment", lambda *a, **k: None
-    )
-    monkeypatch.setattr(
-        "rootstock.commands.manifest.update_and_push_manifest",
+        "rootstock.operations.update_and_push_manifest",
         lambda *a, **k: None,
     )
 
-    from rootstock.commands.install import _install_single_environment
+    from rootstock.operations import install_environment
 
-    rc = _install_single_environment(
+    install_environment(
         root=tmp_path,
         source=str(env_source),
         force=False,
         verbose=False,
+        progress=print,
     )
 
     captured = capsys.readouterr()
-
-    assert rc == 0
     assert "Installing: rootstock==9.9.9" in captured.out
 
     rootstock_install_calls = [
@@ -195,40 +190,32 @@ def test_dependencies_installed_via_uv_sync_script(tmp_path, monkeypatch, capsys
         result.stdout = ""
         return result
 
-    monkeypatch.setattr("rootstock.commands.install.subprocess.run", fake_run)
-    monkeypatch.setattr("rootstock.commands.install.shutil.copy", lambda *a, **k: None)
-    monkeypatch.setattr(
-        "rootstock.commands.install._precompile_environment", lambda *a, **k: None
-    )
+    monkeypatch.setattr("rootstock.operations.subprocess.run", fake_run)
+    monkeypatch.setattr("rootstock.operations.shutil.copy", lambda *a, **k: None)
+    monkeypatch.setattr("rootstock.operations._precompile_environment", lambda *a, **k: None)
     # No vendored wheel available: force the index-spec fallback (also keeps
     # the test off the network).
+    monkeypatch.setattr("rootstock.operations._vendor_rootstock_wheel", lambda root: None)
     monkeypatch.setattr(
-        "rootstock.commands.install._vendor_rootstock_wheel", lambda root: None
-    )
-    monkeypatch.setattr(
-        "rootstock.commands.manifest.update_and_push_manifest",
+        "rootstock.operations.update_and_push_manifest",
         lambda *a, **k: None,
     )
 
-    from rootstock.commands.install import _install_single_environment
+    from rootstock.operations import install_environment
 
-    rc = _install_single_environment(
+    install_environment(
         root=tmp_path,
         source=str(env_source),
         force=False,
         verbose=False,
+        progress=print,
     )
-    assert rc == 0
 
     # Deps install into the staging build dir, which is swapped into envs/
     # only once the whole build succeeds.
     build_dir = str(tmp_path / ".build" / f"withdeps.{os.getpid()}")
 
-    sync_calls = [
-        (cmd, kwargs)
-        for cmd, kwargs in captured_calls
-        if cmd[:2] == ["uv", "sync"]
-    ]
+    sync_calls = [(cmd, kwargs) for cmd, kwargs in captured_calls if cmd[:2] == ["uv", "sync"]]
     assert len(sync_calls) == 1, (
         f"expected exactly one 'uv sync' call, got: {[c for c, _ in captured_calls]!r}"
     )
